more refactor and update API
